requirements.txt
Flask==3.0.0
psycopg[binary]==3.1.13
psycopg-pool==3.2.0
python-dotenv==1.0.0
pytest==7.4.3
openpyxl==3.1.2
reportlab==4.0.7
